Bug fix   @assignment misspelled as @assignments in grades/participants.rhtml
Conversion of multiple mapping/review type to single table designs.
  ReviewMapping -> ResponseMap (subtypes: ParticipantReviewResponseMap, TeamReviewResponseMap)
  ReviewOfReviewMapping -> ResponseMap (subtype: MetareviewResponseMap)
  TeammateReviewMapping -> ResponseMap (subtype: TeammateReviewResponseMap)
  FeedbackMapping -> ResponseMap (subtype: FeedbackResponseMap)
  
  Review, ReviewOfReview, ReviewFeedback, TeammateReview -> Response
  
  Also reorganized Score table to use response_id  
  
  Removed any Response with no scores or no mapping
  Removed any ResponseMap with no response if the assignment it was 
    associated to was older than 1 year
  Removed any Score not associated to a response
  
  Added key constraints to tables to prevent invalid responses/scores in the future.
  
Some benefits include: 
  single View/Controller design for the display of any review type
  cleaner code base
  fewer database tables

  

git-svn-id: https://svn.csc.ncsu.edu:8443/svn/research-expertiza/trunk@628 392d7b7b-3f31-0410-9dc3-c95a9635ea79

+++ app/models/team.rb
@@ -7,7 +7,9 @@ class Team < ActiveRecord::Base
        teamsuser.delete
     end    
     node = TeamNode.find_by_node_object_id(self.id)
-    node.destroy    
+    if node
+      node.destroy
+    end
     self.destroy
   end
